+++ trunk/source/WorldServer/Database/WorldDatabase.h
@@ -28,8 +28,8 @@ public:
 	bool SaveClientLog(const std::string& type, const std::string& message, uint32_t version);
 	uint32_t CreateCharacter(uint32_t account_id, OP_CreateCharacterRequestMsg_Packet* packet);
 	uint16_t GetAppearanceID(std::string name);
-	void UpdateStartingFactions(uint32_t char_id, uint8_t choice);
-	void UpdateStartingZone(uint32_t char_id, uint8_t class_id, uint8_t race_id, uint8_t choice);
+	void UpdateStartingFactions(uint32_t char_id, uint16_t choice);
+	void UpdateStartingZone(uint32_t char_id, uint8_t class_id, uint8_t race_id, uint16_t choice);
 	void UpdateStartingItems(uint32_t char_id, uint8_t class_id, uint8_t race_id, bool base_class = false);
 	void UpdateStartingSkills(uint32_t char_id, uint8_t class_id, uint8_t race_id);
 	void UpdateStartingSpells(uint32_t char_id, uint8_t class_id, uint8_t race_id);
